ILL-1487 add url normalization to prevent invalid caching
This commit adds a utility function, normalize_url, that is
used in _generate_url to sort query parameters in a url
while maintaining the order of lists and duplicate keys.

The purpose of this change is to prevent the situation
where extra parameters are used in a resource lookup,
but may be passed in different ways, causing the same entity
to be cached using more than one url leading
to invalid cache states and cache misses.